Create Physics Calculations
Exercise in using Python for some physics calculations, emulating the sort of things you'd teach in a physics classroom. Some of the print commands required adding in variables, and I got tired of the constant start stop of the + syntax so I looked up how string formatting worked, and it went a lot smoother.