Create Double Index
Exercise:
Create a function named double_index that has two parameters named lst and index.

The function should return a new list where all elements are the same as in lst except for the element at index, which should be double the value of the element at index of lst.

If index is not a valid index, the function should return the original list.

Notes:
Was extremely difficult. Need to remember that the [X:] function will return all values from X index forward in an array. Also need to remember that .append updates without the need for a variable assignment.